requirements.txt
# =====================================================
# Harmony AI - Contoso Analytics Assistant (2025)
# Python 3.9–3.12 Compatible
# =====================================================

# -------------------------
# Core Dependencies
# -------------------------
requests==2.32.3
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1

# -------------------------
# Database (SQL Server)
# -------------------------
pyodbc==5.1.0

# -------------------------
# LLM & Text Processing
# (Ollama installed separately – no pip needed)
# -------------------------
tiktoken==0.7.0       # Token counting
rich==13.7.1          # Better terminal logs

# -------------------------
# Web Frameworks
# -------------------------
streamlit==1.36.0
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-multipart==0.0.9

# -------------------------
# Visualization (minimal)
# (Streamlit built-in charts – no heavy libs needed)
# -------------------------
plotly==5.22.0        # Optional interactive charts

# -------------------------
# Document Processing (optional)
# -------------------------
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
reportlab==4.1.0

# -------------------------
# Utilities
# -------------------------
colorama==0.4.6

# -------------------------
# Testing
# -------------------------
pytest==8.2.2
pytest-asyncio==0.23.6

# -------------------------
# Dev Tools
# -------------------------
black==24.4.2
flake8==7.1.0
mypy==1.11.0
isort==5.13.2
